zsh: implement MVP for per-directory history
Conditional on the activation of the `extended_history` feature, this
feature implementation also adds `histsavecwd` feature. The feature adds
the current directory of the command to each command in the history
file:

    : 1579121354 /home/user:0;ls -l

Instead of the old format:

    : 1579121354:0;ls -l

In addition, the 'fc' commands get a '-c' option that filters out
commands not pertaining to the current directory.

Note that that pathnames that contain the ':' character are not
supported for now - these are expected to mess up history parsing.

Older zsh versions will ignore the added information.